rename format-currency to currency
