+++ apps/web/src/components/currency/index.tsx
@@ -1,7 +1,7 @@
 import { useI18n } from '@/contexts/i18n-context'
 import { useCurrency } from '@/hooks/use-currency'
 import { useLocale } from '@/hooks/use-locale'
-import { formatCurrency, formatIntToFixed } from '@/utils/format-currency'
+import { formatCurrency, formatIntToFixed } from '@/utils/currency'
 
 interface ICurrency {
   value: number
