+++ apps/web/src/components/purchase-form/crypto/sections/crypto-form-instructions.tsx
@@ -9,7 +9,7 @@ import WalletInstructionsModal from '@/components/modals/wallet-instructions'
 import { useI18n } from '@/contexts/i18n-context'
 import { useCurrency } from '@/hooks/use-currency'
 import { useLocale } from '@/hooks/use-locale'
-import { formatCurrency } from '@/utils/format-currency'
+import { formatCurrency } from '@/utils/currency'
 
 export interface CryptoFormInstructionsProps {
   price: number
