+++ apps/web/src/components/purchase-form/wires/sections/bank-account-form.tsx
@@ -16,7 +16,7 @@ import { useI18n } from '@/contexts/i18n-context'
 import { usePaymentContext } from '@/contexts/payment-context'
 import { useCurrency } from '@/hooks/use-currency'
 import { useLocale } from '@/hooks/use-locale'
-import { formatCurrency, formatIntToFixed } from '@/utils/format-currency'
+import { formatCurrency, formatIntToFixed } from '@/utils/currency'
 
 export interface BankAccountFormProps {
   className?: string
