+++ apps/web/src/components/purchase-form/wires/sections/bank-account-success.tsx
@@ -15,8 +15,8 @@ import Heading from '@/components/heading'
 import { useI18n } from '@/contexts/i18n-context'
 import { useCurrency } from '@/hooks/use-currency'
 import { useLocale } from '@/hooks/use-locale'
+import { formatCurrency, formatIntToFixed } from '@/utils/currency'
 import { isAfterNow } from '@/utils/date-time'
-import { formatCurrency, formatIntToFixed } from '@/utils/format-currency'
 import { urls } from '@/utils/urls'
 
 interface BankAccountSuccessProps {
