+++ apps/web/src/components/release-details/sections/release-metadata.tsx
@@ -13,7 +13,7 @@ import Counter from '@/components/counter/counter'
 import { useI18n } from '@/contexts/i18n-context'
 import { useCurrency } from '@/hooks/use-currency'
 import { useLocale } from '@/hooks/use-locale'
-import { formatCurrency } from '@/utils/format-currency'
+import { formatCurrency } from '@/utils/currency'
 
 const { Active, Expired, Upcoming } = PackStatus
 const { Auction, Free, Purchase, Redeem } = PackType
