+++ apps/web/src/templates/my-profile-transactions-template.tsx
@@ -9,7 +9,7 @@ import Pagination from '@/components/pagination/pagination'
 import { useI18n } from '@/contexts/i18n-context'
 import { useCurrency } from '@/hooks/use-currency'
 import { useLocale } from '@/hooks/use-locale'
-import { formatCurrency, formatIntToFixed } from '@/utils/format-currency'
+import { formatCurrency, formatIntToFixed } from '@/utils/currency'
 
 export interface MyProfileTransactionsTemplateProps {
   currentPage: number
