+++ libs/shared/utils/src/index.ts
@@ -2,7 +2,7 @@ export * from './lib/arrays'
 export * from './lib/date-time'
 export * from './lib/encryption'
 export * from './lib/errors'
-export * from './lib/format-currency'
+export * from './lib/currency'
 export * from './lib/http-transport'
 export * from './lib/invariant'
 export * from './lib/logger'
